fix(aic-core): keep run_mixed pass-1 prefix integral (issue #1498)
np.floor returns a float64, so run_mixed's pass-1 RuntimeConfig carried a
float prefix; downstream s = isl - prefix turned float and the DSV4 CSA
CP composition's chunk decomposition (range() over s) raised TypeError —
previously unreachable because the CP path always raised
PerfDataNotAvailableError first (defect 1 of the issue). The product is
integral (prefix x a whole floor count), so int() is lossless.

+++ aic-core/src/aiconfigurator_core/sdk/backends/base_backend.py
@@ -1206,7 +1206,10 @@ class BaseBackend:
                 beam_width=1,
                 isl=num_tokens_combined,
                 osl=1,
-                prefix=prefix * np.floor(step.context_tokens / isl),
+                # int(): np.floor yields a float64 that contaminates the ops'
+                # s = isl - prefix (the DSV4 CP composition chunks s with
+                # range()); the product is integral, so the cast is lossless.
+                prefix=int(prefix * np.floor(step.context_tokens / isl)),
                 seq_imbalance_correction_scale=ctx_scale,
                 engine_step_backend=runtime_config.engine_step_backend,
             ),
